+++ open_mastr/postprocessing/db-cleansing/rli-mastr-solar-cleansing.sql
@@ -11,11 +11,8 @@ __url__     = "https://creativecommons.org/publicdomain/zero/1.0/"
 __author__  = "Ludwig Hülk"
 */
 
-
 ALTER TABLE model_draft.bnetza_mastr_solar_clean
-    ADD COLUMN id SERIAL,
-    ADD PRIMARY KEY (id),
-    ADD COLUMN "comment" text;
+    ADD COLUMN id SERIAL;
 
 ALTER TABLE model_draft.bnetza_mastr_solar_clean
     ALTER COLUMN "Bruttoleistung" TYPE double precision USING "Bruttoleistung"::double precision;